requirements.txt
# MLOps & Versioning
mlflow==2.13.2
dvc[s3]==3.50.1  # Includes base DVC and S3 remote dependencies (for MinIO)

# ML & Data Handling
scikit-learn==1.5.0
pandas==2.2.2

# Configuration & Utilities
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1  # For loading .env files
pyyaml==6.0.1         # For parsing config.yaml
pickle-mixin==1.0.2   # For model serialization
minio==7.2.7          # Python client for MinIO (used in setup script)